Efficient Fitting of (Generalized) Linear Models

Efficient Linear Model ("elm") and Efficient Generalized Linear Model ("eglm") are used to fit linear models in an equivalent way to "lm" and "glm" but in a reduced time depending on the design matrix input (see the DESCRIPTION).

eglm(
  formula,
  family = gaussian(),
  data,
  weights = NULL,
  subset = NULL,
  na.action = na.omit,
  start = NULL,
  etastart = NULL,
  mustart = NULL,
  offset = NULL,
  model = TRUE,
  x = FALSE,
  y = TRUE,
  intercept = TRUE,
  singularity.method = c("eigen", "Cholesky", "qr"),
  tol.solve = .Machine$double.eps,
  tol.values = 1e-07,
  tol.vectors = 1e-07,
  tol.estimation = 1e-08,
  maxit = 25,
  k = 2,
  ...
)

elm(
  formula,
  data,
  subset = NULL,
  weights = NULL,
  na.action = na.omit,
  model = TRUE,
  x = FALSE,
  y = TRUE,
  offset = NULL,
  intercept = TRUE,
  singularity.method = c("eigen", "Cholesky", "qr"),
  tol.solve = .Machine$double.eps,
  tol.values = 1e-07,
  tol.vectors = 1e-07,
  ...
)

Arguments

formula
an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’.

family
a description of the error distribution and link function to be used in the model. For glm this can be a character string naming a family function, a family function or the result of a call to a family function. See family for details of family functions.

data
an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.

weights
an optional vector of weights to be used in the fitting process. Should be NULL or a numeric vector. If non-NULL, weighted least squares is used with weights weights (that is, minimizing sum(w*e^2)); otherwise ordinary least squares is used.

subset
an optional vector specifying a subset of observations to be used in the fitting process.

na.action
a function which indicates what should happen when the data contain NAs. The default is set by the na.action setting of options, and is na.fail if that is unset. The ‘factory-fresh’ default is na.omit. Another possible value is NULL, no action. Value na.exclude can be useful.

start
starting values for the parameters in the linear predictor.

etastart
starting values for the linear predictor.

mustart
starting values for the vector of means.

offset
this can be used to specify an a priori known component to be included in the linear predictor during fitting. This should be NULL or a numeric vector or matrix of extents matching those of the response. One or more offset terms can be included in the formula instead or as well, and if more than one are specified their sum is used. See model.offset.

model
logical value indicating whether model frame should be included as a component of the returned value. Defaults to TRUE.

x, y
logical values indicating whether the model matrix (x) and the response vector (y) used in the fitting process should be returned as components of the returned value. Defaults to FALSE.

intercept
logical value indicating whether intercept should be included in the null model. Defaults to TRUE.

singularity.method
the chosen method to detect for singularity. Defaults to "eigen" but it can also be "Cholesky" or "qr". The spectral decomposition (eigenvectors decomposition) method is slower than Cholesky, but the former accepts tolerance arguments and tends to be more stable for opportune tolerance values.

tol.solve
defaults to .Machine$double.eps, see the function solve.

tol.values
tolerance to consider eigenvalues equal to zero. Defaults to 1e-7.

tol.vectors
tolerance to consider eigenvectors equal to zero. Defaults to 1e-7.

tol.estimation
Tolerance to be used for the estimation. Defaults to 1e-8.

maxit
integer giving the maximal number of IWLS iterations.

k
the penalty per parameter to be used, the default k = 2 is the classical AIC.

...
for elm or eglm: additional arguments to be used to form the default control. On the one hand, use symmetric = F instead of the default TRUE to indicate that the XTX matrix passed internally for least squares estimation is not symmetric. On the other, use out.B = FALSE not to return internal control's matrix output. argument if it is not supplied directly. For weights: further arguments passed to or from other methods.

Value

an object of class "elm" that behaves the same way as the "lm" class, see the function lm.

Details

Models for elm and eglm are specified symbolically. A typical model has the form response ~ terms where response is the (numeric) response vector and terms is a series of terms which specifies a linear predictor for response. A terms specification of the form first + second indicates all the terms in first together with all the terms in second with duplicates removed. A specification of the form first:second indicates the set of terms obtained by taking the interactions of all terms in first with all terms in second. The specification first*second indicates the cross of first and second. This is the same as first + second + first:second, and exactly the same as "lm" and "glm" from the stats package.

Examples

# Linear model
elm(mpg ~ wt, data = mtcars)

#> 
#> Call:
#> elm(formula = mpg ~ wt, data = mtcars)
#> 
#> Coefficients:
#> (Intercept)           wt  
#>      37.285       -5.344  
#> 

# Generalized linear model with Gaussian link
eglm(mpg ~ wt, family = gaussian(), data = mtcars)

#> 
#> Call:  eglm(formula = mpg ~ wt, family = gaussian(), data = mtcars)
#> 
#> Coefficients:
#> (Intercept)           wt  
#>      37.285       -5.344  
#> 
#> Degrees of Freedom: 31 Total (i.e. Null);  30 Residual
#> Null Deviance:	    1126 
#> Residual Deviance: 278.3 	AIC: 166