Predict Method for eglm Fits

Obtains predictions and optionally estimates standard errors of those predictions from a fitted generalized linear model object (see the function predict.glm).

# S3 method for eglm
predict(
  object,
  newdata,
  se.fit = FALSE,
  scale = NULL,
  df = Inf,
  interval = c("none", "confidence", "prediction"),
  level = 0.95,
  type = c("link", "response", "terms"),
  terms = NULL,
  na.action = na.pass,
  weights = NULL,
  ...
)

Arguments

object
a fitted object of class inheriting from "eglm"

newdata
optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted linear predictors are used

se.fit
A switch indicating if standard errors are required.

type
the type of prediction required. The default is on the scale of the linear predictors; the alternative "response" is on the scale of the response variable. Thus for a default binomial model the default predictions are of log-odds (probabilities on logit scale) and type = "response" gives the predicted probabilities. The "terms" option returns a matrix giving the fitted values of each term in the model formula on the linear predictor scale

na.action
function determining what should be done with missing values in newdata. The default is to predict NA

...
further arguments passed to or from other methods

Value

An object of class "numeric" or "matrix", see the function predict.glm.