Fitting Generalized Linear Models

eglm is used to fit generalized linear models in an equivalent way to glm but in a reduced time depending on the design matrix (see the DESCRIPTION).

eglm(
  formula,
  data,
  family = gaussian(),
  intercept = TRUE,
  weights = NULL,
  na.action = na.omit,
  start = NULL,
  etastart = NULL,
  mustart = NULL,
  offset = NULL,
  maxit = 25,
  k = 2,
  model = TRUE,
  singularity.method = c("eigen", "Cholesky", "qr"),
  x = FALSE,
  y = TRUE,
  tol.estimation = 1e-08,
  tol.solve = .Machine$double.eps,
  tol.values = 1e-07,
  tol.vectors = 1e-07,
  ...
)

Arguments

formula
A formula for the model

data
A tibble or data.frame

family
See the function glm, but here it must be specified with brackets (e.g. quasipoisson())

intercept
Logical value to determine wheareas to included an intercept in the null model (Defaults to TRUE)

weights
An optional vector of ‘prior weights’ to be used in the fitting process. Should be NULL or a numeric vector (e.g. data$weights, defaults to NULL)

na.action
a function which indicates what should happen when the data contain NAs. The default is set by the na.action setting of options, and is na.fail if that is unset. The ‘factory-fresh’ default is na.omit. Another possible value is NULL, no action. Value na.exclude can be useful.

start
Starting values for the parameters in the linear predictor

etastart
Starting values for the linear predictor

mustart
Starting values for the vector of means

offset
This can be used to specify an a priori known component to be included in the linear predictor during fitting. This should be NULL or a numeric vector of length equal to the number of cases. One or more offset terms can be included in the formula instead or as well, and if more than one is specified their sum is used. See stats::model.offset()

maxit
See the function glm

k
The penalty per parameter to be used, the default k = 2 is the classical AIC

model
A logical value indicating whether model frame should be included as a component of the returned value (Defaults to TRUE)

singularity.method
The chosen method to detect for singularity. Defaults to "eigen" but it can also be "Cholesky" or "qr"

x
Logical value indicating whether the model matrix used in the fitting process should be returned as components of the returned value (Defaults to FALSE, see the function glm.fit)

y
Logical value indicating whether the response vector used in the fitting process should be returned as components of the returned value (Defaults to FALSE, see the function glm.fit)

tol.estimation
Tolerance to be used for the estimation (Defaults to 1e-8)

tol.solve
(Defaults to .Machine$double.eps, see the function solve)

tol.values
Tolerance to consider eigenvalues equal to zero (Defaults to 1e-7, see the function control),

tol.vectors
Tolerance to consider eigenvectors equal to zero (Defaults to 1e-7, see the function control)

...
For eglm: arguments to be used to form the default control argument if it is not supplied directly. For weights: further arguments passed to or from other methods.

Value

An object of class "eglm" that behaves the same way as the "glm" class, see the function glm.

Examples

# Generalized linear model with Gaussian link
eglm(mpg ~ wt, family = gaussian(), data = mtcars)

#> 
#> Call:  eglm(formula = mpg ~ wt, data = mtcars, family = gaussian())
#> 
#> Coefficients:
#> (Intercept)           wt  
#>      37.285       -5.344  
#> 
#> Degrees of Freedom: 31 Total (i.e. Null);  30 Residual
#> Null Deviance:	    1126 
#> Residual Deviance: 278.3 	AIC: 166